Use client interceptors added in Ratpack 1.6 (#29)
Drops the wrapped HttpClient instrumentation and uses HttpClient Interceptors instead.

This changes the trace results slightly based on the way the Ratpack Interceptor is invoked.  In practice this shouldn't be an issue but if we get complaints we can alway introduce an optional library with the wrapped HttpClient implementation down the road.

By using the Ratpack Interceptor we can close out #8 and it should also fix #27 as well.